# Conflicts:
#	server/application/controllers/Comment_add.php
#	server/application/controllers/Comment_delete.php
#	server/application/controllers/Like_give.php
#	server/application/controllers/Like_isliked.php
#	server/application/controllers/Like_withdraw.php

+++ server/application/models/Likelist_model.php
@@ -5,9 +5,13 @@ class Likelist_model extends CI_Model
     public function __construct()
     {
         $this->load->database();
+        require 'Likelist_constraints.php';
     }
 
     public function getLikesByCommentID($id){
+      if(!isset($id)){
+        throw new Exception('E_PARAM_NOT_EXIST');
+      }
         $result=[];
         $this->db->select('*');
         $this->db->from('LikeList');
@@ -25,17 +29,26 @@ class Likelist_model extends CI_Model
     }
 
     public function deleteLikesByCommentId($cid){
+      if(!isset($cid)){
+        throw new Exception('E_PARAM_NOT_EXIST');
+      }
         $this->db->where('CommentId',$cid);
         $this->db->delete('LikeList');
     }
 
     public function  deleteLikesByCommentIdUserId($cid,$uid){
+      if(!(isset($cid)&&isset($uid))){
+        throw new Exception('E_PARAM_NOT_EXIST');
+      }
         $this->db->where('CommentId',$cid);
         $this->db->where('UserId',$uid);
         $this->db->delete('LikeList');
     }
 
     public function insertLikes($uid,$mid,$cid){
+      if(!(isset($cid)&&isset($uid)&&isset($mid))){
+        throw new Exception('E_PARAM_NOT_EXIST');
+      }
         $data=array(
             'UserId'=>$uid,
             'MusicId'=>$mid,
@@ -44,10 +57,12 @@ class Likelist_model extends CI_Model
         $this->db->insert('LikeList',$data);
     }
 
-    public function isInList($uid,$mid,$cid){
+    public function isInList($uid,$cid){
+      if(!(isset($cid)&&isset($uid))){
+        throw new Exception('E_PARAM_NOT_EXIST');
+      }
         $data=array(
             'UserId'=>$uid,
-            'MusicId'=>$mid,
             'CommentId'=>$cid
         );
         $this->db->select('*');
@@ -55,10 +70,10 @@ class Likelist_model extends CI_Model
         $this->db->where($data);
         $result=$this->db->get();
         if(empty($result->result())){
-            return 1;
+            return false;
         }
         else{
-            return 0;
+            return true;
         }
     }
 }
